Add option for Authorization header in consume tag
This a common authentication method for many of the APIs we deal with and it
would be very convenient if the consume tag supported it. Perhaps this
implementation is too specific (focusing on a single HTTP header field) but if
the need exists, this could be extended to allow for passing additional header
fields.