Fix cookie implementation and use specific cookie for lang
Fix the current implementation for getting and setting the cookie. The
set method was not working. Add a new steam variable 'steam.cookies'
that can be extended in your custom gem to add aditionnal cookies
Change the way to store the selected lang. Instead of storing the
information in the main cookie, use a dedicated cookie.
This allow caching with external tools like varnish.
Adapt the test regarding this change
Add a docker configuration to run the test easily